refactor: consolidate oas responses into models
