+++ virtool/analyses/api.py
@@ -13,7 +13,8 @@ from aiohttp_pydantic import PydanticView
 from aiohttp_pydantic.oas.typing import r200, r204, r400, r403, r404, r409
 from pydantic import conint
 
-from virtool.analyses.oas import AnalysisResponse, FindAnalysesResponse
+from virtool.analyses.models import Analysis
+from virtool.analyses.oas import FindAnalysesResponse
 from virtool.analyses.sql import AnalysisFormat
 from virtool.api.custom_json import datetime_to_isoformat, json_response
 from virtool.api.errors import (
@@ -70,7 +71,7 @@ class AnalysisView(PydanticView):
         self,
         analysis_id: str,
         /,
-    ) -> r200[AnalysisResponse] | r400 | r403 | r404:
+    ) -> r200[Analysis] | r400 | r403 | r404:
         """Get an analysis.
 
         Fetches the details of an analysis.
