+++ virtool/groups/oas.py
@@ -1,6 +1,5 @@
 from pydantic import Field, constr
 
-from virtool.groups.models import Group
 from virtool.models.base import BaseModel
 
 
@@ -40,31 +39,4 @@ class UpdateGroupRequest(BaseModel):
     class Config:
         schema_extra = {
             "example": {"permissions": {"create_ref": True}, "name": "Managers"},
-        }
-
-
-class GroupResponse(Group):
-    class Config:
-        schema_extra = {
-            "example": {
-                "permissions": {
-                    "cancel_job": True,
-                    "create_ref": False,
-                    "create_sample": True,
-                    "modify_hmm": False,
-                    "modify_subtraction": False,
-                    "remove_file": False,
-                    "remove_job": True,
-                    "upload_file": True,
-                },
-                "id": "technicians",
-                "name": "technicians",
-                "users": [
-                    {
-                        "administrator": False,
-                        "handle": "leeashley",
-                        "id": "7CtBo2yG",
-                    },
-                ],
-            },
         }
