+++ virtool/indexes/models.py
@@ -22,6 +22,22 @@ class IndexMinimal(IndexNested):
     reference: ReferenceNested
     user: UserNested
 
+    class Config:
+        schema_extra = {
+            "example": {
+                "change_count": 0,
+                "created_at": "2015-10-06T20:00:00Z",
+                "has_files": True,
+                "id": "fb085f7f",
+                "job": {"id": "bf1b993c"},
+                "modified_otu_count": 0,
+                "ready": False,
+                "reference": {"id": "foo"},
+                "user": {"administrator": False, "handle": "bob", "id": "test"},
+                "version": 9,
+            },
+        }
+
 
 class IndexContributor(UserNested):
     count: int
@@ -47,6 +63,50 @@ class Index(IndexMinimal):
     files: list[IndexFile]
     manifest: dict[str, int]
     otus: list[IndexOTU]
+
+    class Config:
+        schema_extra = {
+            "example": {
+                "version": 0,
+                "created_at": "2015-10-06T20:00:00Z",
+                "ready": False,
+                "has_files": True,
+                "job": {"id": "foo"},
+                "reference": {"id": "foo"},
+                "user": {
+                    "id": "bf1b993c",
+                    "handle": "leeashley",
+                    "administrator": False,
+                },
+                "id": "foo",
+                "change_count": 2,
+                "modified_otu_count": 2,
+                "contributors": [
+                    {
+                        "count": 1,
+                        "id": "fred",
+                    },
+                    {
+                        "count": 3,
+                        "id": "igboyes",
+                    },
+                ],
+                "files": [],
+                "otus": [
+                    {
+                        "change_count": 1,
+                        "id": "kjs8sa99",
+                        "name": "Foo",
+                    },
+                    {
+                        "change_count": 3,
+                        "id": "zxbbvngc",
+                        "name": "Test",
+                    },
+                ],
+                "manifest": [],
+            },
+        }
 
 
 class IndexSearchResult(SearchResult):
