+++ virtool/indexes/oas.py
@@ -1,4 +1,4 @@
-from virtool.indexes.models import IndexSearchResult, IndexMinimal, Index
+from virtool.indexes.models import IndexMinimal, IndexSearchResult
 
 
 class ListIndexesResponse(IndexSearchResult):
@@ -90,50 +90,4 @@ class ReadyIndexesResponse(IndexMinimal):
                     },
                 ]
             }
-        }
-
-
-class GetIndexResponse(Index):
-    class Config:
-        schema_extra = {
-            "example": {
-                "version": 0,
-                "created_at": "2015-10-06T20:00:00Z",
-                "ready": False,
-                "has_files": True,
-                "job": {"id": "foo"},
-                "reference": {"id": "foo"},
-                "user": {
-                    "id": "bf1b993c",
-                    "handle": "leeashley",
-                    "administrator": False,
-                },
-                "id": "foo",
-                "change_count": 2,
-                "modified_otu_count": 2,
-                "contributors": [
-                    {
-                        "count": 1,
-                        "id": "fred",
-                    },
-                    {
-                        "count": 3,
-                        "id": "igboyes",
-                    },
-                ],
-                "files": [],
-                "otus": [
-                    {
-                        "change_count": 1,
-                        "id": "kjs8sa99",
-                        "name": "Foo",
-                    },
-                    {
-                        "change_count": 3,
-                        "id": "zxbbvngc",
-                        "name": "Test",
-                    },
-                ],
-                "manifest": [],
-            },
         }
