+++ virtool/tasks/api.py
@@ -1,16 +1,14 @@
 """Request handlers for querying tasks."""
 
-from typing import List
-
 from aiohttp_pydantic import PydanticView
 from aiohttp_pydantic.oas.typing import r200, r400
 
-from virtool.api.errors import APINotFound
 from virtool.api.custom_json import json_response
+from virtool.api.errors import APINotFound
+from virtool.api.routes import Routes
 from virtool.data.errors import ResourceNotFoundError
 from virtool.data.utils import get_data_from_req
-from virtool.api.routes import Routes
-from virtool.tasks.oas import GetTasksResponse, TaskResponse
+from virtool.tasks.models import Task
 
 routes = Routes()
 
@@ -19,8 +17,7 @@ class TaskServicesRootView(PydanticView):
     """The only endpoint for the task runner and spawner services."""
 
     async def get(self) -> r200:
-        """
-        Root request handler response for task runner.
+        """Root request handler response for task runner.
 
         Used for checking if the server is alive.
 
@@ -38,9 +35,8 @@ class TaskServicesRootView(PydanticView):
 
 @routes.view("/tasks")
 class TasksView(PydanticView):
-    async def get(self) -> r200[List[GetTasksResponse]]:
-        """
-        List all tasks.
+    async def get(self) -> r200[list[Task]]:
+        """List all tasks.
 
         Lists all tasks active on the instance. Pagination is not
         supported.
@@ -53,9 +49,8 @@ class TasksView(PydanticView):
 
 @routes.view("/tasks/{task_id}")
 class TaskView(PydanticView):
-    async def get(self, task_id: int, /) -> r200[TaskResponse] | r400:
-        """
-        Retrieve a task.
+    async def get(self, task_id: int, /) -> r200[Task] | r400:
+        """Retrieve a task.
 
         Fetches the details of a task.
 
@@ -63,7 +58,6 @@ class TaskView(PydanticView):
             200: Successful operation
             404: Not found
         """
-
         try:
             task = await get_data_from_req(self.request).tasks.get(task_id)
         except ResourceNotFoundError:
